app/requirements.txt
# app/requirements.txt
fastapi==0.112.*
uvicorn[standard]==0.30.*
SQLAlchemy==2.0.*
psycopg[binary]==3.2.*
alembic==1.13.*
pydantic==2.*
python-jose[cryptography]==3.3.*
python-multipart==0.0.9
httpx==0.27.*
Faker==25.8.0
python-dotenv==1.0.0
openai==1.43.0
